Moved ICacheGrain
