+++ src/Outkeep.Grains.Interfaces/OutkeepGrainFactoryExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Outkeep.Grains;
+using Outkeep.Grains.Caching;
 using System;
 
 namespace Orleans
