+++ test/Outkeep.Client.Tests/OutkeepDistributedCacheTests.cs
@@ -3,7 +3,6 @@ using Moq;
 using Orleans;
 using Orleans.Concurrency;
 using Outkeep.Core;
-using Outkeep.Grains;
 using Outkeep.Grains.Caching;
 using System;
 using System.Threading.Tasks;
