+++ test/Outkeep.Grains.Interfaces.Tests/GrainFactoryInterfaceExtensionsTests.cs
@@ -1,5 +1,6 @@
 ﻿using Moq;
 using Orleans;
+using Outkeep.Grains.Caching;
 using System;
 using Xunit;
